Update and rename read to README.md
I created a simple model of the cross section of a vein, with red blood cells being procedurally generated. Blue alien cells sometimes appear in the blood and white blood cells then appear to try to catch the alien cells. Each cell type interacts differently, and has a time when it is created and destroyed.